adicionando carrinho - não considerar esse commit

+++ src/main/java/com/desafio/Desafiospring/model/Cart.java
@@ -10,7 +10,9 @@ import java.util.List;
 @Data
 @Builder
 public class Cart {
-    private int id;
+    private long id;
     private List<Product> articles;
     private double total;
 }
+
+
